talk-to-docs/process/requirements.txt
langchain
langchain-community
google-cloud-aiplatform==1.39.0
google-cloud-logging
flask[async]
pandas
pydantic-settings
gcsfs
cloud-sql-python-connector
asyncio==3.4.3 
asyncpg==0.27.0 
pgvector
psycopg2-binary
absl-py==1.4.0
bs4
lxml
trafilatura
unstructured